server/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.12.2
Werkzeug==3.0.4
